+++ hat/assets/js/apps/Iaso/domains/entities/entityTypes/hooks/requests/forms.ts
@@ -4,35 +4,8 @@ import { getRequest } from '../../../../../libs/Api';
 import { useSnackQuery } from '../../../../../libs/apiHooks';
 
 import { usePossibleFields } from '../../../../forms/hooks/useGetPossibleFields';
+import { useGetForm } from '../../../../forms/requests';
 import { Form, PossibleField } from '../../../../forms/types/forms';
-
-export const useGetForm = (
-    formId: number | undefined,
-    enabled: boolean,
-    fields?: string | undefined,
-    appId?: string,
-): UseQueryResult<Form, Error> => {
-    const queryKey: any[] = ['form', formId];
-    let url = `/api/forms/${formId}`;
-    if (fields) {
-        url += `/?fields=${fields}`;
-        if (appId) {
-            url += `&app_id=${appId}`;
-        }
-    } else if (appId) {
-        url += `/?app_id=${appId}`;
-    }
-    return useSnackQuery({
-        queryKey,
-        queryFn: () => getRequest(url),
-        options: {
-            retry: false,
-            enabled,
-            staleTime: 60000,
-            cacheTime: 1000 * 60 * 5,
-        },
-    });
-};
 
 export const useGetForms = (
     enabled: boolean,
